+++ executionUSA/delete.php
@@ -10,31 +10,49 @@ header("Access-Control-Allow-Headers: Content-Type, Access-Control-Allow-Headers
 include_once '../config/database.php';
 include_once '../objects/executionUSA.php';
 
-// instantiate database and product object
+// instantiate database object
 $database = new Database();
 $db = $database->getConnection();
 
 // initialize object
 $executionUSA = new ExecutionUSA($db);
 
-// get id of execution to be edited
-$data = json_decode(file_get_contents("php://input"));
+// get posted data
+if(json_decode(file_get_contents("php://input")) != NULL) {
+	// POST data is JSON format
+	$data = json_decode(file_get_contents("php://input"));
+} else {
+	// POST data is XML format
+	$xmlRequestContents = simplexml_load_string(file_get_contents("php://input"));
+	$xmlInJSON  = json_encode($xmlRequestContents);
+	$data = json_decode($xmlInJSON);
+}
 
 //set ID property of execution to be edited
 $executionUSA->id = $data->id;
 
-if($executionUSA->delete()) {
-	// set response code to 200 (ok)
-	http_response_code(200);
+if(
+	!empty(is_numeric($data->id))
+) {
+	if($executionUSA->delete()) {
+		// Response code - 200 OK
+		http_response_code(200);
 
-	// tell the used the execution was updated
-	echo json_encode(array("message" => "Execution was deleted."));
+		// Show success message
+		echo json_encode(array("message" => "Execution was deleted."));
+	} else {
+		// Response code - 503 SERVICE UNAVAILABLE
+		http_response_code(503);
+
+		// Show error message
+		echo json_encode(array("message" => "Unable to delete execution."));
+	}
 } else {
-	// set response code to 503 (service unavailable)
-	http_response_code(503);
+	// Response code - 400 BAD REQUEST
+	http_response_code(400);
 
-	// tell the used the execution wasn't updated
-	echo json_encode(array("message" => "Unable to delete execution."));
+	// Show error message
+	echo json_encode(array("message" => "Unable to delete execution. Data is incomplete."));
 }
 
 ?>
